pyproject.toml
[tool.poetry]
name = "site-drf"
version = "0.1.0"
description = ""
authors = ["Tatiana_Garbuzova <tatiannna0697@gmail.com>"]
readme = "README.md"
packages = [{include = "site_drf"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
